+++ jax/_src/core.py
@@ -849,11 +849,6 @@ class Tracer(typing.Array, metaclass=StrictABCMeta):
       f"The delete() method was called on {self._error_repr()}."
       f"{self._origin_msg()}")
 
-  def device(self):
-    raise ConcretizationTypeError(self,
-      f"The device() method was called on {self._error_repr()}."
-      f"{self._origin_msg()}")
-
   def devices(self):
     raise ConcretizationTypeError(self,
       f"The devices() method was called on {self._error_repr()}."